analytics-image-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
cloudinary==1.36.0
pillow==10.1.0